pyproject.toml
[project]
name = "fintech-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Семён Верецкий",email = "veretskiisemen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14, <4"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pytest-env (>=1.2.0,<2.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
pytest = "^9.0.2"
pytest-django = "^4.11.1"
pytest-xdist = "^3.8.0"
requests = "^2.32.5"

[tool.black]
line-length = 120
